cast to i8

+++ src/platform_impl/macos/window.rs
@@ -1239,7 +1239,7 @@ impl WindowExtMacOS for UnownedWindow {
 
   #[inline]
   fn set_is_document_edited(&self, edited: bool) {
-    unsafe { self.ns_window.setDocumentEdited_(edited) }
+    unsafe { self.ns_window.setDocumentEdited_(edited as i8) }
   }
 
   #[inline]
